+++ bootstrap/pkgs/default.nix
@@ -1,4 +1,5 @@
 { pkgs }:
 {
   zerolang = pkgs.callPackage ./zerolang.nix { };
+  kernel = pkgs.callPackage ./kernel.nix { };
 }